requirements.txt
aiogram>=3.13,<4.0
aiohttp-socks>=0.9,<1.0
openai>=1.50,<2.0
pydantic-settings>=2.5,<3.0
python-dotenv>=1.0,<2.0
pypdf>=5.0,<6.0
python-docx>=1.1,<2.0
pillow>=10.4,<12.0
pytesseract>=0.3.13,<0.4
pdf2image>=1.17,<2.0
